wifi:  avoid uint8_t in LOG component

+++ src/wifi/model/he-capabilities.cc
@@ -481,9 +481,9 @@ std::ostream &
 operator << (std::ostream &os, const HeCapabilities &HeCapabilities)
 {
   os << HeCapabilities.GetHeMacCapabilitiesInfo1 () << "|"
-     << HeCapabilities.GetHeMacCapabilitiesInfo2 () << "|"
+     << (uint16_t) HeCapabilities.GetHeMacCapabilitiesInfo2 () << "|"
      << HeCapabilities.GetHePhyCapabilitiesInfo1 () << "|"
-     << HeCapabilities.GetHePhyCapabilitiesInfo2 () << "|"
+     << (uint16_t) HeCapabilities.GetHePhyCapabilitiesInfo2 () << "|"
      << HeCapabilities.GetSupportedMcsAndNss ();
   return os;
 }
